fix(vm): keep a string's pool slot at full width on the free list
The string collector pushed each freed slot onto its free list as a u16.
Once the pool passed 65536 strings, a freed slot such as 70000 went back
out as 4464, a slot a live string still held, and the next long string
written there overwrote it. A program holding 70000 strings through a
collection read back 61070 of them wrong.

The free list now holds u32 slot ids, the width the string value itself
carries.

+++ vm/src/ffi.rs
@@ -231,7 +231,7 @@ pub fn c_struct_to_candela_struct(
     struct_fields: &[(SmolStr, DataType, Span)],
     r: &mut RegisterFile,
     recursion_stack: &mut RegisterFile,
-    free_strings: &mut Vec<u16>,
+    free_strings: &mut Vec<u32>,
     gc_string_threshold: &mut u32,
     gc: &mut GcScratch,
     structs: &[Struct],
